Changed all checking of bounds with size() to size_t and removed redundant variables and redundant ; in include statements. Shifted parser class members to header file.

+++ EmptyGeneralTesting/source/DesignExtractor.cpp
@@ -174,9 +174,9 @@ std::vector<int> DesignExtractor::extractCalledByT(vector<vector<int>>col, int s
 vector<vector<int>> DesignExtractor::extractProcModifiesUses(vector<vector<int>>calls, vector<vector<int>> col) {
 	Graph graph(calls.size());
 	
-	for (int i = 0; i < calls.size(); i++) {
+	for (size_t i = 0; i < calls.size(); i++) {
 		vector<int> list = calls.at(i);
-		for (int j = 0; j < list.size(); j++) {
+		for (size_t j = 0; j < list.size(); j++) {
 			graph.addEdge(i, list.at(j));
 		}
 	}
@@ -185,7 +185,7 @@ vector<vector<int>> DesignExtractor::extractProcModifiesUses(vector<vector<int>>
 	if (col.size()==0) {
 	}
 	else {
-		for (int i = 0; i < calls.size(); i++) {
+		for (size_t i = 0; i < calls.size(); i++) {
 			updatedCol = graph.DFS(i);
 		}
 	}
@@ -236,7 +236,7 @@ int DesignExtractor::extractAffectsBothNum(int stmtNum1, int stmtNum2, vector<ve
 	int firstIndex;
 	int secondIndex;
 
-	for (int i = 0; i < startEndNum.size(); i++) {
+	for (size_t i = 0; i < startEndNum.size(); i++) {
 		int start = startEndNum.at(i).first;
 		int end = startEndNum.at(i).second;
 		if ((start <= stmtNum1) && (stmtNum1 <= end)) {
@@ -267,7 +267,7 @@ int DesignExtractor::extractAffectsBothNum(int stmtNum1, int stmtNum2, vector<ve
 
 			for (int i = startNum; i <= endNum; i++) {
 				vector<int> list = nextCol.at(i);
-				for (int j = 0; j < list.size(); j++) {
+				for (size_t j = 0; j < list.size(); j++) {
 					cfg.addEdge(i, list.at(j));
 				}
 			}
@@ -284,7 +284,7 @@ int DesignExtractor::extractAffectsBothNum(int stmtNum1, int stmtNum2, vector<ve
 				vector<int> modifies;
 				int found = 0;
 				if (stmtNum1 != stmtNum2) {
-					for (int i = 1; i < path.size(); i++) {
+					for (size_t i = 1; i < path.size(); i++) {
 						stmtNum = path.at(i);
 						if (stmtNum == stmtNum2) {
 							break;
@@ -319,7 +319,7 @@ int DesignExtractor::extractAffectsBothNum(int stmtNum1, int stmtNum2, vector<ve
 					if (parentT.size() == 0) {
 						return 0;
 					} 
-					for (int i = 0; i < parentT.size(); i++) {
+					for (size_t i = 0; i < parentT.size(); i++) {
 						stmt = parentT.at(i);
 						if (type.at(stmt) == Enum::TYPE::WHILE) {
 							whileStmt.push_back(stmt);
@@ -330,7 +330,7 @@ int DesignExtractor::extractAffectsBothNum(int stmtNum1, int stmtNum2, vector<ve
 						return 0;
 					}
 					int max = whileStmt.at(0);
-					for (int j = 1; j < whileStmt.size(); j++) {
+					for (size_t j = 1; j < whileStmt.size(); j++) {
 						if (whileStmt.at(j) > max) {
 							max = whileStmt.at(j);
 						}
@@ -339,7 +339,7 @@ int DesignExtractor::extractAffectsBothNum(int stmtNum1, int stmtNum2, vector<ve
 					vector<int> children = childrenCol.at(max);
 					
 					vector<int> modified;
-					for (int j = 0; j <children.size(); j++) {
+					for (size_t j = 0; j <children.size(); j++) {
 						if (type.at(children.at(j)) == Enum::TYPE::IF) {
 							continue;
 						}
@@ -375,7 +375,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsFirstNum(int stmtNum1, vec
 	vector<pair<int, int>> results;
 	vector<int> modifies;
 
-	for (int i = 0; i < startEndNum.size(); i++) {
+	for (size_t i = 0; i < startEndNum.size(); i++) {
 		procStart = startEndNum.at(i).first;
 		procEnd = startEndNum.at(i).second;
 		if ((procStart <= stmtNum1) && (stmtNum1 <= procEnd)) {
@@ -386,7 +386,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsFirstNum(int stmtNum1, vec
 	Graph cfg(procEnd+1);
 	for (int i = procStart; i <= procEnd; i++) {
 		vector<int> list = nextCol.at(i);
-		for (int j = 0; j < list.size(); j++) {
+		for (size_t j = 0; j < list.size(); j++) {
 			cfg.addEdge(i, list.at(j));
 		}
 	}
@@ -397,7 +397,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsFirstNum(int stmtNum1, vec
 	int betweenStmt, found = 0;
 
 	vector<int> path = cfg.DFSOriginal(stmtNum1);
-	for (int i = 1; i < path.size(); i++) {
+	for (size_t i = 1; i < path.size(); i++) {
 		stmtNum2 = path.at(i);
 		if (type.at(stmtNum2) == Enum::TYPE::ASSIGN) {
 			usedVar = usesCol.at(stmtNum2);
@@ -406,7 +406,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsFirstNum(int stmtNum1, vec
 					results.push_back(make_pair(stmtNum1, stmtNum2));
 				}
 				else {
-					for (int j = 1; j < i; j++) {
+					for (size_t j = 1; j < i; j++) {
 						betweenStmt = path.at(j);
 						if (type.at(betweenStmt) == Enum::TYPE::WHILE || type.at(betweenStmt) == Enum::TYPE::IF) {
 							continue;
@@ -444,7 +444,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsSecondNum(int stmtNum2, ve
 	vector<pair<int, int>> results;
 	vector<int> modifies;
 
-	for (int i = 0; i < startEndNum.size(); i++) {
+	for (size_t i = 0; i < startEndNum.size(); i++) {
 		procStart = startEndNum.at(i).first;
 		procEnd = startEndNum.at(i).second;
 		if ((procStart <= stmtNum2) && (stmtNum2 <= procEnd)) {
@@ -455,7 +455,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsSecondNum(int stmtNum2, ve
 	Graph cfg(procEnd+1);
 	for (int i = procStart; i <= procEnd; i++) {
 		vector<int> list = prevCol.at(i);
-		for (int j = 0; j < list.size(); j++) {
+		for (size_t j = 0; j < list.size(); j++) {
 			cfg.addEdge(i, list.at(j));
 		}
 	}
@@ -465,7 +465,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsSecondNum(int stmtNum2, ve
 	int modifiedVar;
 	int betweenStmt, found = 0;
 	vector<int> path = cfg.DFSOriginal(stmtNum2);
-	for (int i = 1; i < path.size(); i++) {
+	for (size_t i = 1; i < path.size(); i++) {
 		stmtNum1 = path.at(i);
 			if (type.at(stmtNum1) == Enum::TYPE::ASSIGN) {
 				modifiedVar = modifiesCol.at(stmtNum1).at(0);
@@ -541,7 +541,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsBothUnspecified(vector<vec
 	
 	vector<pair<int, int>> results;
 	vector < pair<int, int>> returnResults;
-	for (int i = 0; i < startEndNum.size(); i++) {
+	for (size_t i = 0; i < startEndNum.size(); i++) {
 
 		int startNum = startEndNum.at(i).first;
 		int endNum = startEndNum.at(i).second;
@@ -552,7 +552,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsBothUnspecified(vector<vec
 			}
 		}
 
-		for (int k = 0; k < returnResults.size(); k++) {
+		for (size_t k = 0; k < returnResults.size(); k++) {
 			results.push_back(returnResults.at(k));
 		}
 	}
@@ -563,7 +563,7 @@ int DesignExtractor::extractAffectsTBothNum(int stmtNum1, int stmtNum2, vector<v
 	int firstIndex;
 	int secondIndex;
 	
-	for (int i = 0; i < startEndNum.size(); i++) {
+	for (size_t i = 0; i < startEndNum.size(); i++) {
 		int start = startEndNum.at(i).first;
 		int end = startEndNum.at(i).second;
 		if ((start <= stmtNum1) && (stmtNum1 <= end)) {
@@ -586,7 +586,7 @@ int DesignExtractor::extractAffectsTBothNum(int stmtNum1, int stmtNum2, vector<v
 		
 		for (int i = startNum; i <= endNum; i++) {
 			vector<int> list = nextCol.at(i);
-			for (int j = 0; j < list.size(); j++) {
+			for (size_t j = 0; j < list.size(); j++) {
 				cfg.addEdge(i, list.at(j));
 			}
 		}
@@ -599,13 +599,13 @@ int DesignExtractor::extractAffectsTBothNum(int stmtNum1, int stmtNum2, vector<v
 			vector<pair<int,int>> affectsIntermediateResults;
 			vector<int> list;
 			vector <vector<int>> affects(endNum+1);
-			for (int i = 0; i < path.size(); i++) {
+			for (size_t i = 0; i < path.size(); i++) {
 					affectsIntermediateResults = extractAffectsFirstNum(path.at(i), modifiesCol, usesCol, nextCol, startEndNum, type, parentTCol, childrenCol);
 					//cout << "Stmt" << path.at(i) << ": ";
-					for (int j = 0; j < affectsIntermediateResults.size(); j++) {
+					for (size_t j = 0; j < affectsIntermediateResults.size(); j++) {
 						list.push_back(affectsIntermediateResults.at(j).second);
 					}
-					for (int k = 0; k < list.size(); k++) {
+					for (size_t k = 0; k < list.size(); k++) {
 						//cout << list.at(k) << " ";
 					}
 					affects.at(path.at(i)) = list;
@@ -615,10 +615,10 @@ int DesignExtractor::extractAffectsTBothNum(int stmtNum1, int stmtNum2, vector<v
 			Graph affectsGraph(endNum+1);
 			int stmt;
 			vector<int> stmts;
-			for (int j = 0; j < path.size(); j++) {
+			for (size_t j = 0; j < path.size(); j++) {
 					stmt = path.at(j);
 					stmts = affects.at(stmt);
-					for (int k = 0; k < stmts.size(); k++) {
+					for (size_t k = 0; k < stmts.size(); k++) {
 						affectsGraph.addEdge(stmt, stmts.at(k));
 					}
 			}
@@ -639,7 +639,7 @@ vector<pair<int,int>> DesignExtractor::extractAffectsTFirstNum(int stmtNum1, vec
 	int procStart, procEnd;
 	vector<int> modifies;
 
-	for (int i = 0; i < startEndNum.size(); i++) {
+	for (size_t i = 0; i < startEndNum.size(); i++) {
 		procStart = startEndNum.at(i).first;
 		procEnd = startEndNum.at(i).second;
 		if ((procStart <= stmtNum1) && (stmtNum1 <= procEnd)) {
@@ -650,14 +650,14 @@ vector<pair<int,int>> DesignExtractor::extractAffectsTFirstNum(int stmtNum1, vec
 	Graph cfg(procEnd + 1);
 	for (int i = procStart; i <= procEnd; i++) {
 		vector<int> list = nextCol.at(i);
-		for (int j = 0; j < list.size(); j++) {
+		for (size_t j = 0; j < list.size(); j++) {
 			cfg.addEdge(i, list.at(j));
 		}
 	}
 
 	vector<int> path = cfg.DFSOriginal(stmtNum1);
 	int check;
-	for (int i = 0; i < path.size(); i++) {
+	for (size_t i = 0; i < path.size(); i++) {
 		check = extractAffectsTBothNum(stmtNum1, path.at(i), modifiesCol, usesCol, nextCol, startEndNum, type, parentTCol, childrenCol);
 		if (check == 1) {
 			results.push_back(make_pair(stmtNum1, path.at(i)));
@@ -672,7 +672,7 @@ vector<pair<int, int>> DesignExtractor::extractAffectsTSecondNum(int stmtNum2, v
 	int procStart, procEnd;
 	vector<int> modifies;
 
-	for (int i = 0; i < startEndNum.size(); i++) {
+	for (size_t i = 0; i < startEndNum.size(); i++) {
 		procStart = startEndNum.at(i).first;
 		procEnd = startEndNum.at(i).second;
 		if ((procStart <= stmtNum2) && (stmtNum2 <= procEnd)) {
@@ -683,14 +683,14 @@ vector<pair<int, int>> DesignExtractor::extractAffectsTSecondNum(int stmtNum2, v
 	Graph cfg(procEnd + 1);
 	for (int i = procStart; i <= procEnd; i++) {
 		vector<int> list = prevCol.at(i);
-		for (int j = 0; j < list.size(); j++) {
+		for (size_t j = 0; j < list.size(); j++) {
 			cfg.addEdge(i, list.at(j));
 		}
 	}
 
 	vector<int> path = cfg.DFSOriginal(stmtNum2);
 	int check;
-	for (int i = 0; i < path.size(); i++) {
+	for (size_t i = 0; i < path.size(); i++) {
 		check = extractAffectsTBothNum(path.at(i), stmtNum2, modifiesCol, usesCol, prevCol, startEndNum, type, parentTCol, childrenCol);
 		if (check == 1) {
 			results.push_back(make_pair(path.at(i), stmtNum2));
@@ -703,7 +703,7 @@ vector<pair<int,int>> DesignExtractor::extractAffectsTBothUnspecified(vector<vec
 	
 	vector<pair<int, int>> results;
 	vector < pair<int, int>> returnResults;
-	for (int i = 0; i < startEndNum.size(); i++) {
+	for (size_t i = 0; i < startEndNum.size(); i++) {
 
 		int startNum = startEndNum.at(i).first;
 		int endNum = startEndNum.at(i).second;
@@ -715,7 +715,7 @@ vector<pair<int,int>> DesignExtractor::extractAffectsTBothUnspecified(vector<vec
 			
 		}
 
-		for (int k = 0; k < returnResults.size(); k++) {
+		for (size_t k = 0; k < returnResults.size(); k++) {
 			results.push_back(returnResults.at(k));
 		}
 	}
